Add additional parameter to file backup

+++ simplyprint_ws_client/helpers/file_backup.py
@@ -8,8 +8,14 @@ class FileBackup:
     """ Small wrapper for count based file backups, used for configs """
 
     @staticmethod
-    def backup_file(file: Path, max_count: int = 5, max_age: Optional[datetime.timedelta] = None):
+    def backup_file(file: Path, max_count: int = 5, min_age_interval: Optional[datetime.timedelta] = None,
+                    max_age: Optional[datetime.timedelta] = None):
         """Backup a file with a count based system
+
+        :param file: The file to back up
+        :param max_count: The maximum number of backups to keep
+        :param min_age_interval: The minimum time between backups
+        :param max_age: The maximum age of a backup
 
         Use the following format
 
@@ -24,13 +30,22 @@ class FileBackup:
         # then adjust the count of the remaining ones
         backups = sorted(file.parent.glob(f"{file.name}.bak.*"), reverse=True)
 
-        if max_age:
-            for backup in backups:
-                date_changed = datetime.datetime.fromtimestamp(backup.stat().st_mtime)
+        latest_backup: Optional[datetime.datetime] = None
 
-                if datetime.datetime.now() - date_changed > max_age:
-                    backup.unlink()
-                    backups.remove(backup)
+        for backup in backups:
+            date_changed = datetime.datetime.fromtimestamp(backup.stat().st_mtime)
+
+            # Keep track of the latest backup
+            if not latest_backup or date_changed > latest_backup:
+                latest_backup = date_changed
+
+            if max_age and datetime.datetime.now() - date_changed > max_age:
+                backup.unlink()
+                backups.remove(backup)
+
+        # If the last backup is too recent, don't create a new one and stop this function
+        if min_age_interval and latest_backup and datetime.datetime.now() - latest_backup < min_age_interval:
+            return
 
         for j, backup in enumerate(backups):
             i = len(backups) - j - 1
